requirements.txt
# amqp==5.0.5
asgiref==3.3.1
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
Babel==2.9.1
boto3==1.18.28
botocore==1.21.28
cachetools==4.2.2
capslock==1.0.5
celery==5.0.5
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.1.5
django-allauth==0.42.0
django-braces==1.14.0
django-cors-headers==3.6.0
django-oauth-toolkit==1.5.0
django-storages==1.11.1
django-timezone-field==4.2.1
drf-yasg==1.20.0
filelock==3.0.12
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==0.9.1
kiwisolver==1.3.1
kombu==5.0.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.20.1
oauthlib==3.1.1
packaging==21.0
pandas==1.2.4
parsel==1.6.0
phpserialize==1.3
Pillow==8.1.0
prompt-toolkit==3.0.16
protobuf==3.17.3
psycopg2-binary==2.8.6
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
service-identity==21.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2
sqlparse==0.4.1
text-unidecode==1.3
tldextract==3.1.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.3
validate-email==1.3
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.12.1
zope.interface==5.4.0
mysqlclient ==2.1.0
uritools ==4.0.0
platformdirs ==2.5.2
dnspython ==2.2.1
urlextract ==1.6.0
tesserocr ==2.5.2
pytesseract ==0.3.9
happytransformer ==2.4.1

nltk ==3.7



